[TASK] Remove dummy files from t3lib
With the switch to namespaces content of most class files were
extracted from t3lib to other extensions like core or backend and
the classes were renamed.

The original files were kept as dummys since 6.0 as a last
convenience layer for extensions that still used require or
require_once of core files manually instead of the class loader
introduced with 4.3 to not make those extensions fatal.

These dummy files are now finally removed.

Be aware that *using* old class names like 't3lib_div' is still
possible with the class_alias mechanism in the class loader, the
only change is now, that manually requiring core files is not
possible anymore. The alias mechanism is *not* removed
with 6.2.

Extension authors should just *remove* require and require_once
calls, the core will take care of finding the correct class name
on instantiation and static usage.

This first patch removes only those files from t3lib, that contain
only the deprecation comment and a require of the new file location.

Change-Id: I874fe47dbd80f43023d0e10686917351f5a1024e
Resolves: #47754
Releases: 6.2
Reviewed-on: https://review.typo3.org/20363
Reviewed-by: Thomas Maroschik
Tested-by: Thomas Maroschik
Reviewed-by: Felix Kopp
Tested-by: Felix Kopp
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn