[BUGFIX] Form localization depending on old class names
Since the introduction of the PHP namespaces some of the localisation of
FORM is not working anymore. The localisation label is depending on the
class name by using get_class(). The localisation is depending on the
old class names which has not been adapted afterwards.

This patch solves this issue by removing the get_class() calls by
using a constant.

Change-Id: I7ba846620732b61e05e210f227c02b63c1198284
Resolves: #47817
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/20415
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>