[BUGFIX] ProcessedFile is persisted in sys_file as well
The problem appears more generally as soon as @getProperties@ is
called from a file object. FAL will try, by all means, indexing the file
(storing the file into sys_file, in other word)

The change set adds a method telling a File not be indexed by the API.
This is particularly useful if you want to create objects but don't want
them to be persisted.

Change-Id: I8cf0c0f6389dc13e8c3855452f1d2519544b69b9
Fixes: #47211
Fixes: #47782
Releases: 6.0, 6.1, 6.2