[BUGFIX] Fix last line in table wizard
In the table wizard (table ce) you have the possibility to move first line
to the last line and vice versa. But on last line the false icon and
input field is set. We also have a FIXME comment there.

With the patch we fix the problem and you can again move the last line on
top of the lines.

Resolves: #16531
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I7fe69397ba595a6b2aec8cdfba05b286f3b8ebe4
Reviewed-on: https://review.typo3.org/20918
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-on: https://review.typo3.org/20920