[BUGFIX] User settings can only save password and text to DB
The user settings module has its own configuration:
http://docs.typo3.org/typo3cms/CoreApiReference/
UserSettingsConfiguration/Columns/Index.html

The property 'table' => 'be_users' allows to save the value of a field
to DB in addition to saving it in the UC array.

This currently only works for types password and text, because only for
those an additional '[be_users]' is added to the form input element
property "name".

Add the additional element to types check and select too.
Additionally convert the check element to a boolean value before handing
it over to TCEmain.

I do not consider this critical enough to port the bug to 4.x.

Resolves: #47966
Releases: 6.2, 6.1, 6.0
Change-Id: Icf69d480013ff1be6c0b5d43212c179468885344