[BUGFIX] Allow colon in displayCond value
Explode the displayCond string to max 4 parts.

This change makes it possible to have colons in the
displayCond value because everything after the third colon
is kept as the fourth part.

Without this change everything after the 4th column (which
could be part of the displayCond value) is ignored.

Resolves: #47988
Releases: 4.7, 4.6, 4.5
Change-Id: Ib32a0ad5cc12e6b98802b18d277ae652dfb45537