[BUGFIX] RTEmagic upgrade wizard fails and uses wrong directory
The RTE upgrade wizard fails, because it uses a
non-existing field name as the table and writes wrong
image paths back into the RTE content. Also the migration
directory should equal the already available migration
directory of the FAL upgrade wizards.

You can test the issue by adding an image to a
RTE field in a 4.7 instance and try to migrate to 6.1
afterwards.

Fixes: #47998
Releases: 6.2, 6.1
Change-Id: Ie6d3947f475e80ac36eb5ba43f258b5215e19928