[!!!][-FEATURE] Drop ext:tsconfig_help
The extension tsconfig_help provides the "TypoScript Help" backend
module and is used by some wizards.

The help information is based on is hopelessly outdated data and
was last updated for CMS 4.1. The included update process parses
manual.sxw files and fatals since 6.0 without anyone noticing. The
manual files are not updated in core for quite some time.

After talking with the doc team, we decided to remove the whole
extension without and all using points now. Any core internal
help mechanics should be based on docs.typo3.org. The doc team
already works on a system to provide current information with the
core, and that will not be based on tsconfig_help.

The patch removes the extension, the wizard entry point, the
wizard register in TCA and other usages without further deprecation.

Change-Id: Ica74245f4925291d8d5ab8b1589420c5abc2be05
Resolves: #47931
Releases: 6.2