[BUGFIX] Adjust extbase record type in TypoScript configuration
During the namespacing in Extbase a mistake happend and the record type
setting in the TypoScript configuration, shipped with
extbase for fe_users and fe_usergroups, got namespaced as well.

This caused in the first place updating instances retrieving
user regords to fail, because existing records had the non namespaced
string set in the database.
Second it rendered newly created records not findable, because those got
the non namespaced string set in database as well.
Now all the records were matched against the namespaced comparison and
failed being considered valid by the repository.

So the solution for both scenarios is to reset the TypoScript config
back to non namespaced string, bringing back the matching constraint
for existing records as well as for newly created ones.

Resolves: #48103
Relates: #40098
Releases: 6.0
Change-Id: Idf02c06b70d337469005a9a6ae773f8cf4d7f03f
Reviewed-on: https://review.typo3.org/20692
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind