[BUGFIX] MailUtility::parseAddresses() returns @localhost on empty string
parseAddresses should return an array of email addresses ready to be
feed into swiftmailer.
If an empty string is passed to the function, then it returns
"@localhost", which is not a valid email.
Swiftmailer will throw an exception "Address in mailbox given
[@localhost] does not comply with RFC 2822, 3.6.2."
This happens in ContentObjectRenderer->sendNotifyEmail() which is in
turn used by the frontend login forgotten passwort form.

Skip any addresses without a mailbox given.

Resolves: #48247
Releases: 6.2, 6.1
Change-Id: If8724416b3451417db7729dc81abdbd068dc8b10
Reviewed-on: https://review.typo3.org/20905
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe