[BUGFIX] compressJs returns wrong filenames
If you enable
    config.concatenateJs = 1
    config.compressJs = 1
then the uncompressed version of the file is linked.

This does not change much, because compression is disabled and the files
will always be uncompressed, however the right link should be generated.

The allows extension to override the compress method and implement
compression without fixing the rest of the code.

The bug is a regression introduced with: #27694 (4.6).

Return the correct filename by overriding the filename with the returned
value of compressJsFile.

Resolves: #48213
Related: #27694
Releases: 6.2, 6.1, 6.0, 4.7
Change-Id: I1f65576db8084b54f9e5a13947ced258eef893d7