[BUGFIX] File upload in RTE-Link-Editor throws 500 error page
When you want to add a link in the RTE-Editor, you can upload files in
the "File" tab. When you select a file and want to upload it, the
reloaded content in the create-link-overlay will throw an error.

The method is feed with a translation of "none - new file".

Bypass the file or folder lockup if we do not have an integer id
and print the message instead.

Resolves: #48219
Releases: 6.2, 6.1, 6.0
Change-Id: I9e1e6bc84b21fb04608841515ac414eec8c04130