[BUGFIX] be_users:userMods is too small to hold all modules
A very long standing bug. It is impossible to assign all modules to a
single user, because the DB field is too small. Saving such a record
will result in a DB error and a truncated modules list.

Solution: Change to field from varchar(255) to text to support a larger
number of modules.

This change is requires a DB compare to become active and does nothing
on its own. Therefore I suggest to also port this to 4.x.

Resolves: #16762
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I67b2ed6b650f9eaf5bdcc9ef81ce93cbbdc54b7f