[BUGFIX] Increase the width of the content element of type Plain HTML
The current editor field for content elements of type HTML is for too
small. Without t3editor it is only 5 lines in height and with t3editor
it is limited to 200px which is not much more.

Increase the default size for tt_content.bodytext to 80x15 and set the
height for t3editor fields to 60% as done in setup.

Resolves: #43112
Releases: 6.2, 6.1, 6.0
Change-Id: Icc27cb013cc3009dace5dedef2f9515af1975ec0
Reviewed-on: https://review.typo3.org/20923
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack
Reviewed-on: https://review.typo3.org/21766