[TASK] Hide func menu with only one entry
The func menu in a module is used to switch between different views in
the same module.

However if only one item is available, there is no reason to show the
function menu at all.
Also the onchange handler does not work as nothing is changeable.

Count the number of elements and suppress the selector if there is only
one item.

Change-Id: I6bbce41af19c8755c67c7ce5085f94930f4f87f6
Resolves: #48175
Releases: 6.2