[BUGFIX] Exceptions if EXT:context_help is not installed
If the extension 'context_help' is not installed, the web
server's error log sometimes shows exceptions like "Class
'TYPO3\CMS\ContextHelp\ExtDirect\ContextHelpDataProvider'
not found" and "Invalid argument supplied for foreach()".
This happens because the class in question is registered
to ExtDirect without checking if the corresponding
extension is loaded.

Change-Id: Ic773cff2822912aac7841955be2f09d77169fdb9
Resolves: #48394
Releases: 6.0, 6.1, 6.2
Reviewed-on: https://review.typo3.org/22293
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn