[BUGFIX] Mysql Socket connection fails
Some people use a socket to connect to a (local) TYPO3 database.
This is not possible any more with TYPO3 6.1, because it is not possible
any more to specify a socket.
Formerly one could append the socket to the hostname similar to the port
setting.

Introduce a new config option for the socket.
$GLOBALS['TYPO3_CONF_VARS']['DB']['socket']

The default socket is NULL, thus the behavior is unchanged if no socket
is defined.

@see http://php.net/manual/en/mysqli.real-connect.php

The installer is adapted to handle the socket setting and a bit of
JavaScript magic is used to hide unrelated fields.

Resolves: #48117
Releases: 6.2, 6.1
Change-Id: I968a72dbaf90b3f319c03e697137630534d325c1