[BUGFIX] Support mysql socket connection
With the switch to mysqli in 6.1, the connection to mysql server via
socket broke.
With mysqli, the definition is now:
* If hostname is "localhost", a socket is used and the socket file
can be given optionally if it is different from the default.
* If hostname is not "localhost" but an IP address, a port can be
given that may be different from the default mysql 3306.

This backport just considers the LocalConfiguration.php setting and
does not change anything in the installer.

Resolves: #48117
Releases: 6.2, 6.1
Change-Id: I968a72dbaf90b3f319c03e697137630534d325c1
Reviewed-on: https://review.typo3.org/21973
Reviewed-by: Markus Klein
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn