[BUGFIX] Do not add category to empty table names
If the $GLOBALS['TYPO3_CONF_VARS']['SYS']['defaultCategorizedTables']
is empty or you have empty strings between commas then we try to add
empty tablenames to the table definitions.

This is fixed by setting removeEmptyValues to true in trimExplode
and adding a RuntimeException to the add function to prevent
others (extensions) breaking the functionality.

Resolves: #48381
Releases: 6.2, 6.1, 6.0
Change-Id: I4a8d43a0ff75da298ace4f1d33a5dcf2d2ac569f