[BUGFIX] Filelist: Only show paste buttons if clipObj is set
During introduction of the paste buttons for empty folders, I assumed
that the clipboard object is always set.
This might not be the case, e.g. in EXT:sphinx 4.0.0.

Wrap the paste button code with a check if the clibObj is set.

We can not set the clibObj here, because it depends on variables of the
calling class.

Resolves: #48603
Releases: 6.2, 6.1, 6.0
Change-Id: I2c7c322c9cb994cfdfa1511c96e8c7da143dc77a