[BUGFIX] findByUid should respect query settings of repository
Prior to 6.1 findByUid built its own query which respected
the settings of its repository. Now findByUid just fetches
the object from the persistence manager which builds its own
query without respecting the query setting of the repository.

This leads to the error that the user ins't able to fetch
hidden or deleted records by uid any more. Therefore the
old behaviour of the repository must be used.

Releases: 6.2, 6.1
Resolves: #48571
Change-Id: Ic9a48d5d06f20b2ab8032c39eecb89cd52f898af