[BUGFIX] findObjectByIdentifier in Backend must not respect enable fields
With the introduction of enhanced query settings the method
findObjectByIdentifier changed its behaviour as enable fields
are respected by default which is wrong for this method.

The methods functionality is defined by fetching an object by
its identifier which is not marked as deleted. Therefore the
only constraint must be "deleted = 0".

Releases: 6.2, 6.1
Resolves: #48690
Change-Id: I9d13dfd5fc205fb7f262c8d09bbf14ac136e9203