[BUGFIX] FormFieldVH->getPropertyValue() must work without form object
Since the new property mapper is activa by default it is
possibe to create forms without object="{object}", e.g. in
a newAction. This leads to some ViewHelpers throwing an
Exception as they rely on a present object to recieve the
property value. As this object may not be present a missing
check hast to be added to method getPropertyValue().

This check also exists in Flow.Fluid and is necessary to
fix malfunctioning view helpers.

Releases: 6.2, 6.1
Resolves: #48750
Relates: #48628
Change-Id: I6c42725d6daf9003358f8b6fbe9af3e62f011213