[BUGFIX] FormFieldVH->getPropertyValue() must work without form object
Since the new property mapper is active by default it is
possible to create forms without object="{object}", e.g.
using a newAction. This leads to some view helpers throwing
an exception as they rely on a present object to receive the
property value. As this object may not be present a missing
check has to be added to method getPropertyValue().

This check also exists in Flow.Fluid and is necessary to
fix malfunctioning view helpers.

Releases: 6.2, 6.1
Resolves: #48750
Relates: #48628
Change-Id: I6c42725d6daf9003358f8b6fbe9af3e62f011213
Reviewed-on: https://review.typo3.org/21181
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Alexander Schnitzler
Tested-by: Alexander Schnitzler
Reviewed-on: https://review.typo3.org/21220