[BUGFIX] Enable CheckboxViewHelper binding to ArrayObject and Null
Currently the checkbox view helper relies on a present
form object which is not necessary any longer since 6.1
as the new property mapper is active by default.

Therefore the viewhelper must be adjusted to be able to
be bound to ArrayObject objects or Null which throws an
exception till now.

Releases: 6.2, 6.1
Resolves: #48628
Change-Id: I35b70fb0a09ddd72da1c64ec5c7c5456e9af9471