[BUGFIX] OpenID login does not work with Google
The claimed ID may differ from the URL to access the OpenID provider;
a notable example is Google:
Their OpenID is https://www.google.com/accounts/o8/id for all users, while
the actual claimed ID is different, e.g.
https://www.google.com/accounts/o8/id?id=AItOawm2w
(for users without a google+ profile).

Because of this, we cannot verify that the OpenID URL given by the user
on login exists in our database before starting the OpenID process.
Instead, we first need to run the OpenID process to get the final
claimed identifier and only then check if a user with that ID exists
in our database.

This patch moves the start of the OpenID process from authorize()
into getUser().

It also introduces a new field in the backend login form, "openid_url".
The field is needed to cleanly distinguish OpenID from normal logins,
to make full OpenID URL normalization possible (i.e. adding http://
automatically if missing).

Resolves: #25322
Releases: 6.1, 6.2
Change-Id: Id31238760fe4e58e2a823beddaeb454ba28d59be