[BUGFIX] Non accessible pages on shortcuts and PageNotFound handler
The var pageNotFound is set, if the called page has access restrictions.
Afterwards starts a searching for an accessible page in the rootline upwards.

If that page is a short link which also isn't accessible we stop instead of
searching again in this new rootline. Limiting this to a maximum of 20
iterations to prevent endless loops.

If an accessible page is found we do not reset the pageNotFound var. The
PageNotFound handler reacts on this var and redirects to the 404 page
instead of presenting the accessible page we found later on.

You can reproduce this with the introduction package, for example change the
access to the Example/Tables page to "Customer". Afterwards go to
http://yourdomain/?id=38 and you will see the 404 page. If you disable the
pageNotFound_handling you will see the content of the Example page.

Resolves: #16472
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I1e58ec1f96422c6bf3e5c9c74f1b1c1666b68762