[BUGFIX] EM TimeSinceLastUpdateViewHelper fatal if never updated
If the repository was never updated (or the update failed) last_update
have the value of zero in the database.

The function
\TYPO3\CMS\Extbase\Persistence\Generic\Mapper\DataMapper:mapDateTime()
will return NULL (instead of a DateTime object for 1.1.1970 00:00:00),
so we try NULL->format in the TimeSinceLastUpdateViewHelper and this
leads to the fatal error.

Reproduce:
In the DB set last_updated column to 0 in for the first (and only) entry
in table tx_extensionmanager_domain_model_repository, then go into the
"Get Extensions" menu. Now a JSON request is fired to get last update time
which will return HTTP 500 Internal Server Error (you can see also the
flash message for that error).

Resolves: #49131
Releases: 6.2, 6.1, 6.0
Change-Id: Ia41c50f2037fb1c7ef935a8855a200b303347383