[BUGFIX] Web->Page - Language view not usable
Since the "all language" elements are displayed in each
language (introduced in 4.7), the language view of the
page module is not really usable anymore. If you have
activated the mod.web_layout.defLangBinding setting in the
Page TS configuration  "all language" elements are not
displayed anymore. Also you cannot overwrite "all language"
elements in an easy way inside another languages
and overwritten elements are displayed together with their
"all language" default element inside a column. Furthermore
the sorting order is sometimes broken on non-default languages
if the "all language" element is rendered inside the column.
The patch fixes all this issues in the languages view by
reverting the changes introduced the 4.7 branch just for the
languages view. The columns view works like it was intented with
the fix from the 4.7 branch (it still has the same flaws as
the languages view, but this issues can be fixed with another
changeset).

Fixes: #42768
Fixes: #42771
Releases: 6.2, 6.1, 6.0
Change-Id: Ibd222c3980ebcc7a5f273d85b3d99d6938291e9e
Reviewed-on: https://review.typo3.org/21891
Reviewed-by: Stefan Galinski
Tested-by: Stefan Galinski