[BUGFIX] Web->Page - Language view not usable
Since the "all language" elements are displayed in each
language (introduced in 4.7), the language view of the
page module is not really usable anymore. If you have
activated the defLangBinding setting "all language"
elements are not displayed anymore. Also you cannot
overwrite "all language" elements in an easy way anymore
inside another language and overwritten elements are
displayed together with the default element inside a
column. Furthermore the sorting order is sometimes
broken.

The patch doesn't fixes the described issues in the columns
function, because it's simply not relevant for multilanguage
sites and must be treated in another changeset if required.

Fixes: #42768
Fixes: #42771
Releases: 6.2, 6.1, 6.0
Change-Id: Ibd222c3980ebcc7a5f273d85b3d99d6938291e9e