[BUGFIX] Shortcuts within mounted trees lose MP parameter
When mounting a tree which has shortcut pages in it, a link to such a
shortcut page has the MP parameter in it. But when redirecting to the
destination page of the shortcut, this MP parameter is lost, thus making
the page lose its connection to the tree root (or rather to the part of
the rootline above the mountpoint).

Change-Id: If438927a005e00c7f0128606cebd5091e95f5e17
Fixes: #17537
Branches: 4.5, 6.0, 6.1, 6.2