[BUGFIX] Flexform field is available but not passed to sub-function calls
This change fixes three occurrences of calls to FlexForm DS
processing methods which accept a $fieldName argument but in
the current state of the code is given an empty string instead
of the $field variable which exists in all three contexts and
contains the proper value.

Passing the field name along to these sub functions allows hook
subscribers to use the field name - among other things. In any
case this fixes an unexpected behaviour turning it into an
expected one.

Resolves: #49331
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I078b0734f3bfa74a9acd61773c5c380f091286f8