[TASK] Allow template paths to use overlays
This change allows the TemplateView to use TypoScript configured
overlay paths which will be checked for the existence of the
template file which is expanded from the path pattern. Instead of
checking just one path, the class now checks all path sets defined
in `plugin.tx_extkey.view.overlays.someidorextkey.templateRootPath`
for example.

NB: this is the implementation that has existed in EXT:view for
quite a while and has made many integrators very happy. It has
been tested and found very useful as a pattern but is indeed
different from overriding just one template file.

Fixes: #39868 although not by using fallbacks but by letting the
current behaviour be used as a sort of fallback and other paths
overlaid instead.

Documentation and a short how-to can be copy/pasted freely from:
https://github.com/FluidTYPO3/view#template-path-overlays

Releases: 6.0, 6.1, 6.2
Change-Id: Ib3fb37c8642a8259f98dcf0493986ee43a341472