[!!!][FEATURE] Allow template paths to use overlays
This change allows the TemplateView to use TypoScript configured
overlay paths which will be checked for the existence of the
template file which is expanded from the path pattern. Instead of
checking just one path, the class now checks all path sets defined
in `plugin.tx_extkey.view.overlays.someidorextkey.templateRootPath`
for example.

This way it is now possible to override the location of single
templates, partials and layouts, by just putting the template in
question into some directory and registering the directory as
overlay. When fluid looks up the location of a single template
or partial, it will take the overlay directories into account and
will use a template from an overlay if it is found, instead of the
original resource.

Fixes: #39868 although not by using fallbacks but by letting the
current behaviour be used as a sort of fallback and other paths
overlaid instead.

It should be noted that Sebastian Kurfuerst from Flow.Fluid is
still against such a feature in FLOW3, but practical reasons show
there is a general demand to have such an overlay system in CMS
fluid. We hope that Flow.Fluid in the future also implements a
similar solution, even if it introduces complexity.

Documentation and a short how-to can be copy/pasted freely from:
https://github.com/FluidTYPO3/view#template-path-overlays

Resolves: #39868
Releases: 6.2
Change-Id: Ib3fb37c8642a8259f98dcf0493986ee43a341472