[TASK] Remove last bit of multiplyDBfieldSize handling
multiplyDBfieldSize was used in the past to extend the size of
database fields to cope with UTF8 content if the database did
not support it. The setting was deprecated for a long time already
and removed from DefaultConfiguration with version 6.0.
The patch removes still existing handler code from ext:install.
The setting is also removed from LocalConfiguration if found.

Change-Id: Ia36b8c05cab8814c91ffa0d8e8afcd4de5c7060b
Resolves: #49334
Related: #35923
Related: #49299
Releases: 6.2