[BUGFIX] Cannot load css_styled_content TS using INCLUDE_TYPOSCRIPT
When installing extensions they may register plugins in ext_localconf.php.
By doing so some default TypoScript is generated in
Extensionutility::configurePlugin() that is then added using
ExtensionManagementUtility::addTypoScript().

Parts of the generated TypoScript are stored in
$TYPO3_CONF_VARS['FE']['contentRenderingTemplates']* and are loaded later
during TypoScript parsing after including the content rendering template.

These TS parts are used to resolve which classes and methods to call when
a plugin needs to be rendered on a page. You can find them in
tt_content.list.20.*

"content default" used to be the old content rendering template with uid
43, today we use css_styled_content instead. The part in the TypoScript
processing chain to load these generated snippets -
TemplateService::includeStaticTypoScriptSources() - is only executed when
loading the content rendering templates using the "Include static (from
extensions)" field in template records.

In return this means if you try to move as much TS to files as possible
you will still have to include css_styled_content through the Template
records in the backend. In other words when actually trying to include
css_styled_content TS from a file, the generated snippets for
tt_content.list.20.* will not be loaded and thus no plugin can be
executed.

Resolves: #49461
Releases: 6.2, 6.1, 6.0
Change-Id: I77c0d64759fa0b5b27cac0ae1acd55e9abefe7f8