[BUGFIX] Fix property mapping of checkbox arrays
If multiple checkboxes are rendered but none of it is
selected, an exception is thrown as the value of the hidden field,
that is submitted is a string.

This must be the case because we need to be able to clear out
selected checkboxes and cannot craft a hidden field that will
be translated to a GET/POST arument which is an empty array.

Because of that the ArrayConverter must be able to convert
empty strings to an empty array, just like the "old"
property mapper did.

Change-Id: I115831014caca4868d1f421eb968e01178e1169b
Resolves: #47832
Releases: 6.2, 6.1, 6.0
Reviewed-on: https://review.typo3.org/21700
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring