[BUGFIX] Do not show System Extensions Wizard when done
The check if the wizard has been executed before and
includes all available system extensions have been installed
is wrong. Rather than casting to bool, we should
check if the difference between "seen extensions" and
available system extensions is an empty array.

This correctly shows the wizard when not finished and
hides it when finished.

Resolves: #49256
Releases: 6.2, 6.1, 6.0
Change-Id: I7bd655b2b7eeaaa5b1dd91e80564f150b1d0d038