[BUGFIX] use additional HTTP Header to determine if request was sent via HTTPS
The HTTP Header "X-Forwarded-Proto" is a defacto standard which may be
sent by Frontend-Proxies/Load-Balancers to let its Backend-Server
know which protocol was used to start the original request.

SSL Requests can be terminated at a transparent Frontend-Proxy
(e.g. nginx) and passed to the Backend-Server via plain HTTP in
this case.

Change-Id: Ie6ab10f503737b2146032911b2587e1e96e9db0f