[BUGFIX] RTE: Importing google webfonts breaks style sheet parsing
Problem: Firefox raises a security error because the file is not in the
same domain as the executing script. Probably a Firefox bug since the
file certainly has adequate headers: the font are indeed displayed in
the RTE.

Solution: Since the Webfont style sheet does not contain any css rules
anyways, we will catch the error, and if it is a security error, log it
to the Javascript console (or Firebug) and silently continue. Re-throw
any other type of error.

Resolves: #49648
Releases: 4.5, 4.7, 6.0, 6.1, 6.2
Change-Id: I225152ad5d84de50f29d3830be6f1f3e10f249c8