[FEATURE] Enable double-resolution icons in SpriteGenerator
In order to display high resolution icons to according displays
the SpriteManager must be enabled to support this.

This patch adds support for that. It generates a second sprite
with double sized icons and uses background-size CSS3 property.
As this sprite is a separate file it does not affect performance
on non high Resolution Displays. High Resolution displays directly
load the second file - which indeed is 4 times the size.

For generating the high resolution sprite all icons which are not
available as @x2-variant are scaled with gd-library functionality.

Change-Id: Id606f0c2900d57bb46869829d5f0939f29e9db08
Resolves: #49667
Documentation: #52853
Releases: 6.2
Reviewed-on: https://review.typo3.org/21887
Reviewed-by: Christian Kuhn
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Tested-by: Christian Kuhn
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter