[FEATURE] Add meta info to the FAL IRRE element header
Implement the UX-Team functionility of the FAL IRRE element header via a
label_userFunc in the sys_file_reference. Furthermore add a
configuration in the ctrl-section of the sys_file_reference TCA to
support additional fields in the output of the header.
This can be used by any extension.

Resolves: #36052
Releases: 6.2
Change-Id: I7770044c11a6c1a1bcb51a3bfc11b85923653196
Reviewed-on: https://review.typo3.org/21916
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter