[FEATURE] Add checkFlexFormValue hook in DataHandler
Currently there is a problem with updating flexform values
right as there is no check if an existing value in the db
has actually been deleted.

This e.g. is a problem when it comes to saving switchable
controller actions (Extbase) in the flexform. Old actions
that may not be present any longer have to be removed
manually.

As there is currently no overall solution to this issue
there should at least be a hook to adjust the old flexform
data right before it is merged with the new one.

Releases: 6.2
Resolves: #49699
Change-Id: Ic93ffd7782b4d049070c6bd9ab760a01008242d6