[BUGFIX] Change the label userfunc for FAL records
The new label_userFunc for sys_file_reference rows results in
"no title" when called by BackendUtility::getRecordTitle. This function
is e.g. called from the workspaces module.

This patch will set the title parameter to the row uid when the
parmeter array does not provide an 'options' 'sys_file' array. This
will result in the uid being displayed which is the old behavior.

Resolves: #49718
Related: #36052
Releases: 6.2
Change-Id: Ib340ca82f5b10faa00815a074545664c2fa67954